refactor: 데모데이 투표 요청·응답 DTO를 표준 패키지로 이동
